Share app version across mobile apps
