+++ mobile/apps/auth/lib/ui/settings_page.dart
@@ -8,7 +8,6 @@ import 'package:ente_auth/store/code_store.dart';
 import 'package:ente_auth/ui/components/buttons/button_widget.dart';
 import 'package:ente_auth/ui/settings/about_settings_page.dart';
 import 'package:ente_auth/ui/settings/account_settings_page.dart';
-import 'package:ente_auth/ui/settings/app_version_widget.dart';
 import 'package:ente_auth/ui/settings/components/auth_settings_navigation.dart';
 import 'package:ente_auth/ui/settings/components/auth_settings_page_scaffold.dart';
 import 'package:ente_auth/ui/settings/data/data_settings_page.dart';
@@ -25,6 +24,7 @@ import 'package:ente_components/ente_components.dart';
 import 'package:ente_lock_screen/local_authentication_service.dart';
 import 'package:ente_strings/ente_strings.dart';
 import 'package:ente_ui/components/settings/app_engagement_section.dart';
+import 'package:ente_ui/components/settings/app_version_widget.dart';
 import 'package:ente_ui/components/settings/social_icons_row.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
