+++ mobile/packages/ui/lib/components/settings/app_version_widget.dart
@@ -1,0 +1,45 @@
+import 'package:ente_components/ente_components.dart';
+import 'package:ente_strings/ente_strings.dart';
+import 'package:flutter/material.dart';
+import 'package:package_info_plus/package_info_plus.dart';
+
+class AppVersionWidget extends StatefulWidget {
+  const AppVersionWidget({super.key});
+
+  @override
+  State<AppVersionWidget> createState() => _AppVersionWidgetState();
+}
+
+class _AppVersionWidgetState extends State<AppVersionWidget> {
+  late final Future<PackageInfo> _packageInfo;
+
+  @override
+  void initState() {
+    super.initState();
+    _packageInfo = PackageInfo.fromPlatform();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return FutureBuilder<PackageInfo>(
+      future: _packageInfo,
+      builder: (context, snapshot) {
+        if (!snapshot.hasData) {
+          return const SizedBox.shrink();
+        }
+
+        return Padding(
+          padding: const EdgeInsets.symmetric(vertical: Spacing.xl),
+          child: Center(
+            child: Text(
+              context.strings.appVersion(versionValue: snapshot.data!.version),
+              style: TextStyles.mini.copyWith(
+                color: context.componentColors.textLight,
+              ),
+            ),
+          ),
+        );
+      },
+    );
+  }
+}
