+++ mobile/packages/ui/pubspec.yaml
@@ -26,6 +26,7 @@ dependencies:
   fluttertoast: 8.2.14
   hugeicons: 1.1.7
   modal_bottom_sheet: 3.0.0
+  package_info_plus: 8.3.1
   url_launcher: 6.3.2
 
 dev_dependencies:
